requirements.txt
langgraph>=0.2.0
langchain-groq>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
rank-bm25>=0.2.2
python-dotenv>=1.0.0
